+++ run_tests.sh
@@ -1,5 +1,4 @@
-#!/usr/bin/env sh
-rm **/*.cov
+#!/usr/bin/env zsh
 julia --project=@. -e 'using Pkg; Pkg.test(coverage=true)'
 julia --project=coverage coverage/produce_coverage.jl
 rm **/*.cov