+++ test/test_field.jl
@@ -9,22 +9,22 @@
 
     for step_size in (1., 0.3, 0.123)
         step_size = Float(step_size)
-        @testset "Simulate empty environment/sequence with varying step sizes" begin
+        @testset "Simulate empty environment/sequence with dt=$step_size" begin
             orient = mr.evolve_to_time(mr.Spin(transverse=1.), zero(Float), one(Float), mr.Microstructure(), step_size).orientations[1]
             @test mr.phase(orient) == 0.
         end
-        @testset "Test constant off-resonance field" begin
+        @testset "Test constant off-resonance field with dt=$step_size" begin
             orient = mr.evolve_to_time(mr.Spin(transverse=1.), zero(Float), Float(0.3), mr.Microstructure(off_resonance=mr.field(2.)), step_size).orientations[1]
-            @test mr.phase(orient) ≈ mr.norm_angle(mr.rad2deg(0.3 * 2. * 3 * mr.gyromagnetic_ratio))
+            @test mr.phase(orient) ≈ mr.norm_angle(0.3 * 2. * 360)
         end
-        @testset "Test gradient off-resonance field" begin
+        @testset "Test gradient off-resonance field with dt=$step_size" begin
             micro = mr.Microstructure(off_resonance=mr.field(SA[1.5, 0., 0.], 2.))
             spin = mr.evolve_to_time(mr.Spin(transverse=1.), zero(Float), Float(0.3), micro, step_size)
             orient = spin.orientations[1]
-            @test mr.phase(orient) ≈ mr.norm_angle(mr.rad2deg(0.6 * 3 * mr.gyromagnetic_ratio))
+            @test mr.phase(orient) ≈ mr.norm_angle(0.6 * 360)
             # Move spin and evolve a bit further in time
             orient = mr.evolve_to_time(mr.Spin(transverse=1., position=SA[3., 0., 0.]), zero(Float), Float(0.2), micro, step_size).orientations[1]
-            @test mr.phase(orient) ≈ mr.norm_angle(mr.rad2deg(((2. + 1.5 * 3.) * 0.2) * 3 * mr.gyromagnetic_ratio))
+            @test mr.phase(orient) ≈ mr.norm_angle((2. + 1.5 * 3.) * 0.2 * 360)
         end
     end
     @testset "Fields with different types" begin